Change Buffer to write bits to the end of scratch and read from the start
This solves the problem of sending the buffer over the network.
This is because we can only send bytes over the network. So we can
have upto 7 bits of "excess" data. This then screws up the calculations
when we try to shift right to get the data back out. instead we end up
with the extra padding zeros instead of the bits we actually want.